# swagger/swagger.yaml
openapi: 3.1.0
info:
  title: CTScan Company
  description: Lorem Ipsum
  version: 1.0.12
servers:
  - url: http://localho.st:8080
    description: Local server
tags:
  - name: Authentication
    description: Authentication and Authorization
  - name: UserService
    description: UserAccountService
  - name: AnalysisService
    description: Analyzation of CTScan and Reporting
  - name: Payment
    description: Payment Service
paths:
  /login:
    post:
      tags:
        - Authentication
      summary: Login user.
      description: This can only be done by the logged in user.
      operationId: loginUser
      requestBody:
        description: Login user
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: The name for the user.
                  example: john doe
                password:
                  type: string
                  description: The password for the user.
                  example: password123
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          description: Invalid username/password supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /logout:
    post:
      tags:
        - Authentication
      summary: Logout user.
      description: This can only be done by the logged in user.
      operationId: logoutUser
      requestBody:
        description: Logout user
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  description: The unique identifier for the user.
                  example: 1
        required: true
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid username/password supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account:
    post:
      tags:
          - UserService
      summary: Create user account.
      description:
      requestBody:
        description: Create a new user account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: user created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /account/{userId}:
    get:
      tags:
        - UserService
      summary: Get user account by userId.
      description: Get user account detail based on userId.
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: The userId that needs to be fetched.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid userId supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /upload:
    post:
      tags:
        - AnalysisService
      summary: Uploads CT Scan.
      description: Uploads CT Scan
      operationId: uploadCTScan
      requestBody:
        description: Upload CT Scan
        content:
          application/json:
            schema:
              type: object
              properties:
                file:
                  type: string
                  description: Base64 encoded string of the CT Scan file
                  example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
        required: true
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid CT Scan supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /analyse:
      post:
        tags:
          - AnalysisService
        summary: Analyse CT Scan.
        description: Analyse CT Scan
        operationId: analyseCTScan
        requestBody:
          description: Analyse CT Scan
          content:
            application/json:
              schema:
                type: object
                properties:
                  ctScanId:
                    type: integer
                    description: Scan
                    example: "1"
                  owner:
                    type: integer
                    description: User Id
                    example: "1"
          required: true
        responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/CtScanAnalysis'
          '400':
            description: Invalid CT Scan supplied
          default:
            description: Unexpected error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/Error"
          '404':
            description: CT Scan not found
  /report/{reportId}:
    get:
      tags:
        - AnalysisService
      summary: Get report.
      description: Get report
      operationId: getReport
      parameters:
        - name: reportId
          in: path
          description: The reportId that needs to be fetched.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          description: Invalid reportId supplied
        '404':
          description: Report not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /report:
    post:
      tags:
        - AnalysisService
      summary: Create report.
      description: Create report
      operationId: createReport
      requestBody:
        description: Create report
        content:
            application/json:
              schema:
                type: object
                properties:
                  ctScanAnalysisId:
                    type: integer
                    description: Scan
                    example: "1"
                  content:
                    type: string
                    description: Report content
                    example: "<html><body><h1>CT Scan Report</h1></body></html>"
                  owner:
                    type: integer
                    description: User Id
                    example: "1"
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          description: Invalid Report supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: CT Scan Analysis not found
  /payment:
    post:
      tags:
        - Payment
      summary: Payment.
      description: Put new payment in the database
      operationId: payment
      requestBody:
        description: Payment
        content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                    description: The unique identifier for the user.
                    example: 1
                  amount:
                    type: number
                    format: integer
                    description: Amount to be paid
                    example: 100.50
        required: true
      responses:
        '200':
          description: successful operation
        '400':
          description: Invalid payment details supplied
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"  
  /payment/{userId}:
    get:
      tags:
        - Payment
      summary: Get payment information.
      description: Get payment information about payment from userId.
      operationId: getPayment
      parameters:
        - name: userId
          in: path
          description: The userId that needs to be fetched.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '400':
          description: Invalid userId supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    User:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: The unique identifier for the user.
        username:
          type: string
          description: The name for the user.
          example: john doe
        email:
          type: string
          example: john@email.com
        type:
          type: int
          description: The type of user. 1 --> Doctor, 2 --> User
          example: 2
          default: 2
        password:
              type: string
              description: The password for the user.
              example: password123
        iban:
          type: string
          description: The IBAN for the user.
          example: NL89370400440532013000
    Session:
      type: object
      properties:
        sessionId:
          type: string
          description: The unique identifier for the session.
          format: uuid
          example: 1234567890abcdef
        userId:
          type: integer
          format: int64
          description: The unique identifier for the user.
          example: 1
        expiryTime:
          type: string
          format: date-time
          description: Datetime of expiry
    Payment:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier for the payment.
          example: 1
        userId:
          type: integer
          format: int64
          description: The unique identifier for the user.
          example: 1
        amount:
          type: number
          format: float
          description: Amount to be paid
          example: 100.50
        createdAt:
          type: string
          format: date-time
          description: Datetime of payment
          example: 2023-10-01T12:00:00Z
    Report:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        createdAt:
          type: string
          format: date-time
          description: Datetime
        content:
          type: string
          format: html
          description: htmlContent
        owner:
          type: integer
          format: int64
          description: UserId
          example: 1
        ctScanAnalysis:
          type: array
          description: Array of CTScan
          items: 
            $ref: "#/components/schemas/CtScanAnalysis"
    CtScan:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        image:
          type: string
          description: Image Blob
          format: binary
        createdAt:
          type: string
          description: Datetime
          format: date-time
        owner:
          type: integer
          format: int64
          description: User Id
          example: 1
    CtScanAnalysis:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        createdAt:
          type: string
          description: Datetime
          format: date-time
        ctScan:
          type: array
          description: One CTScan
          items: 
            $ref: "#/components/schemas/CtScan"
        score:
          type: number
          format: float
          description: Score of the analysis
        owner:
          type: integer
          format: int64
          description: User Id
          example: 1 
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message